refactor the x86 module
