+++ src/x86/i386/mod.rs
@@ -1,0 +1,4 @@
+//! `i386` intrinsics
+
+mod eflags;
+pub use self::eflags::*;
