+++ src/x86/i586/sse.rs
@@ -3,7 +3,7 @@
 use simd_llvm::simd_shuffle4;
 use v128::*;
 use v64::f32x2;
-use super::c_void;
+use x86::c_void;
 use std::mem;
 use std::ptr;
 
@@ -626,23 +626,6 @@ pub unsafe fn _mm_cvt_ss2si(a: f32x4) -> i32 {
     _mm_cvtss_si32(a)
 }
 
-/// Convert the lowest 32 bit float in the input vector to a 64 bit integer.
-///
-/// The result is rounded according to the current rounding mode. If the result
-/// cannot be represented as a 64 bit integer the result will be
-/// `0x8000_0000_0000_0000` (`std::i64::MIN`) or trigger an invalid operation
-/// floating point exception if unmasked (see
-/// [`_mm_setcsr`](fn._mm_setcsr.html)).
-///
-/// This corresponds to the `CVTSS2SI` instruction (with 64 bit output).
-#[inline(always)]
-#[target_feature = "+sse"]
-#[cfg_attr(test, assert_instr(cvtss2si))]
-#[cfg(target_arch = "x86_64")]
-pub unsafe fn _mm_cvtss_si64(a: f32x4) -> i64 {
-    cvtss2si64(a)
-}
-
 // Blocked by https://github.com/rust-lang-nursery/stdsimd/issues/74
 // pub unsafe fn _mm_cvtps_pi32(a: f32x4) -> i32x2
 // pub unsafe fn _mm_cvt_ps2pi(a: f32x4) -> i32x2 { _mm_cvtps_pi32(a) }
@@ -670,23 +653,6 @@ pub unsafe fn _mm_cvttss_si32(a: f32x4) -> i32 {
 #[cfg_attr(test, assert_instr(cvttss2si))]
 pub unsafe fn _mm_cvtt_ss2si(a: f32x4) -> i32 {
     _mm_cvttss_si32(a)
-}
-
-/// Convert the lowest 32 bit float in the input vector to a 64 bit integer
-/// with truncation.
-///
-/// The result is rounded always using truncation (round towards zero). If the
-/// result cannot be represented as a 64 bit integer the result will be
-/// `0x8000_0000_0000_0000` (`std::i64::MIN`) or an invalid operation floating
-/// point exception if unmasked (see [`_mm_setcsr`](fn._mm_setcsr.html)).
-///
-/// This corresponds to the `CVTTSS2SI` instruction (with 64 bit output).
-#[inline(always)]
-#[target_feature = "+sse"]
-#[cfg_attr(test, assert_instr(cvttss2si))]
-#[cfg(target_arch = "x86_64")]
-pub unsafe fn _mm_cvttss_si64(a: f32x4) -> i64 {
-    cvttss2si64(a)
 }
 
 // Blocked by https://github.com/rust-lang-nursery/stdsimd/issues/74
@@ -722,20 +688,6 @@ pub unsafe fn _mm_cvtsi32_ss(a: f32x4, b: i32) -> f32x4 {
 #[cfg_attr(all(test, not(target_os = "macos")), assert_instr(cvtsi2ss))]
 pub unsafe fn _mm_cvt_si2ss(a: f32x4, b: i32) -> f32x4 {
     _mm_cvtsi32_ss(a, b)
-}
-
-/// Convert a 64 bit integer to a 32 bit float. The result vector is the input
-/// vector `a` with the lowest 32 bit float replaced by the converted integer.
-///
-/// This intrinsic corresponds to the `CVTSI2SS` instruction (with 64 bit
-/// input).
-#[inline(always)]
-#[target_feature = "+sse"]
-#[cfg_attr(all(test, target_os = "macos"), assert_instr(cvtsi2ssq))]
-#[cfg_attr(all(test, not(target_os = "macos")), assert_instr(cvtsi2ss))]
-#[cfg(target_arch = "x86_64")]
-pub unsafe fn _mm_cvtsi64_ss(a: f32x4, b: i64) -> f32x4 {
-    a.replace(0, b as f32)
 }
 
 // Blocked by https://github.com/rust-lang-nursery/stdsimd/issues/74
@@ -1133,14 +1085,16 @@ pub unsafe fn _mm_loadr_ps(p: *const f32) -> f32x4 {
 #[cfg_attr(all(test, any(target_arch = "x86_64", target_feature = "sse2")),
            assert_instr(movhpd))]
 pub unsafe fn _mm_storeh_pi(p: *mut u64, a: f32x4) {
-    if cfg!(target_arch = "x86") {
+    #[cfg(target_arch = "x86")]
+    {
         // If this is a `f64x2` then on i586, LLVM generates fldl & fstpl which
         // is just silly
         let a64: u64x2 = mem::transmute(a);
         let a_hi = a64.extract(1);
         *p = a_hi;
-    } else {
-        // target_arch = "x86_64"
+    }
+    #[cfg(target_arch = "x86_64")]
+    {
         // If this is a `u64x2` LLVM generates a pshufd + movq, but we really
         // want a a MOVHPD or MOVHPS here.
         let a64: f64x2 = mem::transmute(a);
@@ -1164,14 +1118,16 @@ pub unsafe fn _mm_storeh_pi(p: *mut u64, a: f32x4) {
                target_family = "windows"),
            assert_instr(movsd))]
 pub unsafe fn _mm_storel_pi(p: *mut u64, a: f32x4) {
-    if cfg!(target_arch = "x86") {
+    #[cfg(target_arch = "x86")]
+    {
         // Same as for _mm_storeh_pi: i586 code gen would use floating point
         // stack.
         let a64: u64x2 = mem::transmute(a);
         let a_hi = a64.extract(0);
         *p = a_hi;
-    } else {
-        // target_arch = "x86_64"
+    }
+    #[cfg(target_arch = "x86_64")]
+    {
         let a64: f64x2 = mem::transmute(a);
         let a_hi = a64.extract(0);
         *p = mem::transmute(a_hi);
@@ -1718,14 +1674,8 @@ extern "C" {
     fn ucomineq_ss(a: f32x4, b: f32x4) -> i32;
     #[link_name = "llvm.x86.sse.cvtss2si"]
     fn cvtss2si(a: f32x4) -> i32;
-    #[link_name = "llvm.x86.sse.cvtss2si64"]
-    #[cfg(target_arch = "x86_64")]
-    fn cvtss2si64(a: f32x4) -> i64;
     #[link_name = "llvm.x86.sse.cvttss2si"]
     fn cvttss2si(a: f32x4) -> i32;
-    #[link_name = "llvm.x86.sse.cvttss2si64"]
-    #[cfg(target_arch = "x86_64")]
-    fn cvttss2si64(a: f32x4) -> i64;
     #[link_name = "llvm.x86.sse.sfence"]
     fn sfence();
     #[link_name = "llvm.x86.sse.stmxcsr"]
@@ -1741,7 +1691,7 @@ extern "C" {
 #[cfg(test)]
 mod tests {
     use v128::*;
-    use x86::sse;
+    use x86::i586::sse;
     use stdsimd_test::simd_test;
     use test::black_box; // Used to inhibit constant-folding.
 
@@ -2860,38 +2810,6 @@ mod tests {
     }
 
     #[simd_test = "sse"]
-    #[cfg(target_arch = "x86_64")]
-    unsafe fn _mm_cvtss_si64() {
-        use std::f32::NAN;
-        use std::i64::MIN;
-        let inputs = &[
-            (42.0f32, 42i64),
-            (-31.4, -31),
-            (-33.5, -34),
-            (-34.5, -34),
-            (4.0e10, 40_000_000_000),
-            (4.0e-10, 0),
-            (NAN, MIN),
-            (2147483500.1, 2147483520),
-            (9.223371e18, 9223370937343148032),
-        ];
-        for i in 0..inputs.len() {
-            let (xi, e) = inputs[i];
-            let x = f32x4::new(xi, 1.0, 3.0, 4.0);
-            let r = sse::_mm_cvtss_si64(x);
-            assert_eq!(
-                e,
-                r,
-                "TestCase #{} _mm_cvtss_si64({:?}) = {}, expected: {}",
-                i,
-                x,
-                r,
-                e
-            );
-        }
-    }
-
-    #[simd_test = "sse"]
     unsafe fn _mm_cvttss_si32() {
         use std::f32::NAN;
         use std::i32::MIN;
@@ -2924,41 +2842,6 @@ mod tests {
     }
 
     #[simd_test = "sse"]
-    #[cfg(target_arch = "x86_64")]
-    unsafe fn _mm_cvttss_si64() {
-        use std::f32::NAN;
-        use std::i64::MIN;
-        let inputs = &[
-            (42.0f32, 42i64),
-            (-31.4, -31),
-            (-33.5, -33),
-            (-34.5, -34),
-            (10.999, 10),
-            (-5.99, -5),
-            (4.0e10, 40_000_000_000),
-            (4.0e-10, 0),
-            (NAN, MIN),
-            (2147483500.1, 2147483520),
-            (9.223371e18, 9223370937343148032),
-            (9.223372e18, MIN),
-        ];
-        for i in 0..inputs.len() {
-            let (xi, e) = inputs[i];
-            let x = f32x4::new(xi, 1.0, 3.0, 4.0);
-            let r = sse::_mm_cvttss_si64(x);
-            assert_eq!(
-                e,
-                r,
-                "TestCase #{} _mm_cvttss_si64({:?}) = {}, expected: {}",
-                i,
-                x,
-                r,
-                e
-            );
-        }
-    }
-
-    #[simd_test = "sse"]
     pub unsafe fn _mm_cvtsi32_ss() {
         let inputs = &[
             (4555i32, 4555.0f32),
@@ -2976,36 +2859,6 @@ mod tests {
                 e,
                 r,
                 "TestCase #{} _mm_cvtsi32_ss({:?}, {}) = {:?}, expected: {:?}",
-                i,
-                a,
-                x,
-                r,
-                e
-            );
-        }
-    }
-
-    #[simd_test = "sse"]
-    #[cfg(target_arch = "x86_64")]
-    pub unsafe fn _mm_cvtsi64_ss() {
-        let inputs = &[
-            (4555i64, 4555.0f32),
-            (322223333, 322223330.0),
-            (-432, -432.0),
-            (-322223333, -322223330.0),
-            (9223372036854775807, 9.223372e18),
-            (-9223372036854775808, -9.223372e18),
-        ];
-
-        for i in 0..inputs.len() {
-            let (x, f) = inputs[i];
-            let a = f32x4::new(5.0, 6.0, 7.0, 8.0);
-            let r = sse::_mm_cvtsi64_ss(a, x);
-            let e = a.replace(0, f);
-            assert_eq!(
-                e,
-                r,
-                "TestCase #{} _mm_cvtsi64_ss({:?}, {}) = {:?}, expected: {:?}",
                 i,
                 a,
                 x,
