+++ src/x86/i686/mod.rs
@@ -1,0 +1,10 @@
+//! `i686` intrinsics
+
+mod sse2;
+pub use self::sse2::*;
+
+mod sse41;
+pub use self::sse41::*;
+
+mod sse42;
+pub use self::sse42::*;
