+++ src/x86/x86_64/mod.rs
@@ -1,0 +1,10 @@
+//! `x86_64` intrinsics
+
+mod sse;
+pub use self::sse::*;
+
+mod sse2;
+pub use self::sse2::*;
+
+mod sse42;
+pub use self::sse42::*;
